📐 alternate solution
with defining defaults outside the trait, i could reduce the number of
errors thrown by the compiler, and get the code working for the most
part, other than for the implementation of `Self` in update() and for
`sleep` field in `blink()`.

still not sure why defining this within the trait isn't working.